+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `isInteger` validation to `number` type form elements
 - `minLength` and `maxLength` validation to `text` and `textarea` type form elements
 - Support for `includeTimestampWatermark` for `camera` form element type
+- sending userToken with requests when present in querystring
 
 ## 0.1.4 (2020-11-12)